IMPALA-7970 : Add support for metastore event based automatic invalidate
This change adds support to CatalogD to poll metastore events to issue
invalidate on tables automatically. It adds basic infrastructure to poll
HMS notifications events at a configurable frequency using a backend
config called hms_event_polling_frequency_s flag. Currently, it issues
invalidate at tables when it received alter events on table and
partitions. It also adds tables/databases and removes tables from
catalogD when it receives create_table/create_database and
drop_table/drop_database events. The default value of
hms_event_polling_frequency_s is 0 which disables the feature. A
non-zero value in seconds of this configuration can be used to enable
the feature and set the polling frequency.

The change also adds a new test class to test the basic functionality
for each of the event type which is supported.

Change-Id: Ic70b27780560b7ac9b33418d132b36cd0ca4abf7