IMPALA-8101: Thrift 11 and ext-data-source compilation are always run
Compilation of Thrift 11 Python code (IMPALA-7924) and of ext-data-source
Thrift files (ErrorCodes.thrift, ExternalDataSource.thrift, Data.thrift,
Status.thrift, Types.thrift) is run during every build, regardless of
whether or not the .thrift files have changed. The issue is that the
CMake custom command for compilation of these files points to a
non-existent OUTPUT_FILE.

This patch fixes Thrift 11 compilation by pointing the OUTPUT_FILE of
each .thrift file to its corresponding __init__.py file. For
compilation of ext-data-source, things are a bit tricky as we only run
Java gen and it is difficult to map Java generated code to the
corresponding .thrift files purely based on file names. Instead, for
ext-data-source, this patch adds a dummy file under
ext-data-source/api/target/tmp/generated-sources/ to track if a .thrift
file has been compiled or not. A `mvn clean` of ext-data-source will
delete all of these files and trigger re-compilation of the
ext-data-source files.

Change-Id: I52520e4b099c7bac5d088b4ba5d8a335495f727d
Reviewed-on: http://gerrit.cloudera.org:8080/12290
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>