IMPALA-8168: Disable Sentry HDFS sync on S3
This patch updates the build script to disable Sentry HDFS sync plugin
on S3 since Sentry HDFS sync is not supported and it may cause issues,
such as making the tests flaky in a non-supported environment.

This patch also updates the CDH_BUILD_NUMBER to 909265 to test the new
Sentry changes for batch insert of HMS paths in full snapshot. It also
bumps Kudu version to 1.10.0.

Testing:
- Ran core tests on both HDFS and S3.

Change-Id: Ifa12f163a32bb6e597323149fbe565a6f66f5e69
Reviewed-on: http://gerrit.cloudera.org:8080/12482
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>