IMPALA-6897: Catalog web-ui exposes top-n tables with most num of files.
Add functions in CatalogUsageMonitor to monitor and report the catalog
usage of the tables have the most number of files
List the tables with the most number of files in Catalog server web-ui
and sorted by the files number.

Testing:
Add tests to check table usage information is in the catalogd's
catalog page.

Change-Id: I04df5756641bb10dbb86d813b8001c4a04d7dc9b
Reviewed-on: http://gerrit.cloudera.org:8080/12483
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>