CDH-77790: Dynamically figure out Kudu (and other CDH) versions.
Similar to CDH-72729, we ran into issues with needing to update Impala
for a Kudu version bump. This uses a properties file that CDH is
publishing to figure out the versions.

The alternative would have been to strip versions from the tarball files
themselves, but I think this is a bit more coherent.

Some of the environment variables were removed because
they're not used. Upstream uses them for impala-parent/pom.xml,
but downstream we use the cdh root pom.

I tested this by running test-with-docker on this branch as well
as a private Cauldron build.

Change-Id: I355796180a1bad99cf279c3ef212baef5ecc5a18
Reviewed-on: https://gerrit.sjc.cloudera.com/41787
Reviewed-by: Hector Acosta <hector.acosta@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>