IMPALA-8222: disable per-query timeouts in stress test
It is very hard to pick a timeout threshold for queries
in the stress test that won't result in false positives,
because the slowdown can be non-linear with the amount
of load on the system.

To avoid this problem this change simply disables the
timeout for the stress test phase. The timeout logic
is still used for query cancellation and generating
random queries (since those may run for too long).

Testing:
Manually tested against my minicluster:
* TPC-H binary search for one query
* A short stress test run
* Random query generation with a 1 second timeout.

Change-Id: I2ee8b8ec63562031784c2a719869dce57bcafc0b
Reviewed-on: http://gerrit.cloudera.org:8080/12531
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>