IMPALA-7979: Fix bug of SkipValues() and GetNextValues() interaction
Alternating calls to DictDecoder's SkipValues() and GetNextValues()
could result in a decoding error.

RleBatchDecoder<T>::DecodeLiteralValues() returned false when
invoked with num_literals_to_consume=0 and had buffered literals.
Then the caller (DictDecoder::GetNextValues()) thought it failed,
and in the end it resulted in a decoding error.

It could only occur if GetNextValues() was invoked after SkipValues().
Otherwise the RleBatchDecoder in DictDecoder could not have buffered
literals, since they were buffered in the dictionary decoder.

Since SkipValues() is not used yet, it couldn't happen in production
code.

I added a backend test to fuzz test the interaction between
SkipValues() and GetNextValues(). I looped the test for a few hours
without seeing an error.

Change-Id: Id0426cdef54ee9bc2306a542098a92c640dc41c4
Reviewed-on: http://gerrit.cloudera.org:8080/12555
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>