IMPALA-8069: Crash in impala::Sorter::Run::Run
Modifies the planner to omit a sorter node in the obscure case we have
an analytic query with a constant sort. The analytic node requires a
buffered tuple. When the sort is omitted, creates a buffered tuple on
top of the node that would have been input to the sort.

Testing:
* Added a PlannerTest case.
* Added an end-to-end test.
* Reran the query that previously crashed: it now produces proper results.

Change-Id: I19a57f3791ce9e41fe0ba79dc5834e762341c74e
Reviewed-on: http://gerrit.cloudera.org:8080/12562
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>