IMPALA-8247: Fix tests missing from unified backend executable
When converting the tests in be/src/util to the unified backend
test executable in IMPALA-8071, some tests did not get linked
into the executable appropriately. This means that they stopped
running. Specifically, the tests in bit-stream-utils-test.cc
and system-state-info-test.cc were left out.

This adds them back and modifies bin/validate-unified-backend-filters.py
to check both directions:
 - If a filter is specified, it must match a test in the executable.
 - If the executable has a test, some filter must match it.
I ran the backend tests on debug and ASAN with this change.

Change-Id: I80d8b5779cb0ac663e32c72722e0c73b68a43d2e
Reviewed-on: http://gerrit.cloudera.org:8080/12584
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>