IMPALA-8249: End-to-end test framework doesn't read aggregated counters properly
Updated compute_aggregation() function to not read the pretty-printed
value from the runtime profile, but the accurate value which is at
the end of the line in parenthesis, e.g.:

RowsReturned: 2.14M (2142543)

The old regex tried to parse '2.14M' with 'd+', which resulted in '2'
instead of 2142543.

I tested the change manually and added a test case to
'tests/unittests/test_result_verifier.py'.

Change-Id: I2a6fc0d3f7cbaa87aa848cdafffad21fb1514930
Reviewed-on: http://gerrit.cloudera.org:8080/12589
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>