IMPALA-8252: Only write node_metadata if it's set
Since IMPALA-1048 we write TRuntimeProfileNode.node_metadata
unconditionally, even when both its fields are unset. This trips up the
Thrift library Java reader code, which expects to find exactly one type
of a union to be set.

This change fixes the issue by checking whether at least one of the
fields is set before marking node_metadata as set.

This change also adds a test to runtime-profile-test.cc to make sure
that the __isset flag is set correctly. Additionally I wrote a profile
and then manually read it with a Java client.

Change-Id: Ice1efcf75daba5c023566eb63d4e921c228c26fd
Reviewed-on: http://gerrit.cloudera.org:8080/12616
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>