IMPALA-8256: Better error message for ImpalaServicePool::RejectTooBusy()
An incoming request to a RPC service can be rejected due to either
exceeding the memory limit or maximum allowed queue length.
It's unclear from the current error message which of those factors
contributes to the failure as neither the actual queue length nor
the memory consumption is printed.

This patch fixes the problem by printing the estimated queue length
and memory consumption when a RPC request is dropped.

Testing done: verified the new error message with test_rpc_timeout.py

Change-Id: If0297658acf2b23823dcb7d2bdff5d8e4475bb98
Reviewed-on: http://gerrit.cloudera.org:8080/12624
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>