IMPALA-8278 : Fix testEventProcessorFetchAfterHMSRestart
MetastoreEventsProcessorTest#testEventProcessorFetchAfterHMSRestart
causes test flakiness because the event processor instantiated is not
stopped. This causes the test to have 2 instances of EventsProcessor
running concurrently during the test execution. Depending on the timing
of the poll operations all the events are processed twice and they
modify the state of test catalog concurrently causing race conditions
which lead to intermittent test failures.

Testing done:
1. Confirmed that two event processors are running by manually adding
identifiers to each event processing and log them when they receive the
events.
2. Ran the test multiple times locally to confirm that test works
everytime.
3. Running full-tests using jenkins job.

Change-Id: Ia12b1d048d8c8c59fa4c587bf65772ff194d314d