IMPALA-8305: Generate JUnitXML for DCHECK failures
DCHECKs log at the FATAL level and produce a statement starting
with "Check failed:". This generates a JUnitXML when it sees a
FATAL log file with "Check failed:". Some backend death tests
specifically test hitting certain DCHECKs, so the be_tests
directory is ignored.

Something that hits a DCHECK would also generate a minidump,
so this is just additional information beyond the existing
JUnitXML for the minidump.

Testing:
 - Tested on logs from a test that hit a DCHECK in impalad
 - Ran a normal error-free run

Change-Id: Ic45d01b6a87b1d81967f1e10028af96a7cc3071a
Reviewed-on: http://gerrit.cloudera.org:8080/12751
Reviewed-by: David Knupp <dknupp@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>