IMPALA-8320: Handle psutil error in tests
We use psutil.Process.cmdline to access a process's command line. This
field is a property and tries to retrieve process information from the
OS. If the process has terminated in the meantime, such access will
result in an error which we need to handle. This change adds proper
error handling. I ran the existing Breakpad tests to make sure we don't
regress, but there's no obvious way to test that particular corner case
that would be worth the effort.

Change-Id: Ia5fec061f845d3deb6328e52e590a7a79196dfd4
Reviewed-on: http://gerrit.cloudera.org:8080/12792
Reviewed-by: Lars Volker <lv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>