IMPALA-7981: Add host disk statistics to profile
This change adds host disk statistics to profiles. For each host that
participates in the query execution it adds the read and write bandwidth
across all disks. This includes all data read or written by the host as
part of the execution of a query (spilling), by the HDFS data node, and
by other processes running on the same system.

The change adds tests for the added functionality to the backend and end
to end tests.

Change-Id: I373e7da47a0d722938e6ca1572c49a502951ed57
Reviewed-on: http://gerrit.cloudera.org:8080/12822
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>