Fix stress test tracking of past metrics values
The stress test keeps track of various metrics about queries that it
has already run, such as the number of completed queries and the
number of errors. The stress test also forks off several processes to
handle running of queries.

The intention is that the metrics should be shared across all of the
processes. However, a recent patch (IMPALA-6662) changed these metrics
to be stored in a dict and in the process caused them to no longer be
shared. This patch solves the problem by storing the value in the dict
as 'Value' objects, which are shared.

Synchronization of access to these values is handled by taking
'_query_runners_lock'.

Testing:
- Ran the stress test for 1000 queries on tpch 500 on a 10 node
  cluster. Previously, this would reach a point where there were 0
  'Active' queries and hang.

Change-Id: I03b5e24b41fff1d33f9bb64b14a6d0b7a23a56e2
Reviewed-on: http://gerrit.cloudera.org:8080/12903
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>