Improve error handling for validation of unified backend executable
bin/validate-unified-backend-test-filters.py currently does not
handle the case where the call to run the unified backend test
executable fails. Instead, it produces a very confusing message
about invalid filters that does not mention that the execution
failed.

This checks the return code when running the unified backend
test executable and produces a reasonable message if the
return code is not 0.

Change-Id: Ia84cd074be79bc9d99b0621a1ae216f3debf5833
Reviewed-on: http://gerrit.cloudera.org:8080/12895
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>