IMPALA-8363: Deny access when column masking or row filtering is enabled in Ranger
This patch updates the Ranger authorization checker code to deny access
when column masking and row filtering is enabled in Ranger for queries
that that have columns/tables specified in column mask and row filter
policies. This is to prevent data leak, such that the data that is
masked/filtered in Hive should not be visible at all in Impala until
Impala has full support for column masking and row filtering.

Testing:
- Added tests in AuthorizationStmtTest to test queries with column
  masking and row filtering enabled.
- Ran all FE tests
- Ran all E2E tests

Change-Id: If46b4bf24d916e4a4ea8a36ff4acfd95d5f45c8e
Reviewed-on: http://gerrit.cloudera.org:8080/12927
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>