Clean up generation of XML configuration files
hive-site.xml and sentry-site.xml in particular had grown multiple
slightly-different variants, differing only in a few small pieces. This
was difficult to maintain: in fact, while attempting to clean them up I
found a number of places that the MySQL and Postgres versions of
hive-site had diverged for no apparent reason.

This moves away from using the sed-based templating for these
configuration files, and instead uses python as a poor man's template
system. That enables much simpler conditional logic.

I briefly considered XSLT for this, but decided that Python is probably
easier for the average developer to follow, modify, and debug.

Along the way, I removed a few flags which appear to be no longer used
by Hive 2 or later, and a few items which were already commented out in
the previous template:

- hive.stats.dbclass
- hive.stats.dbconnectionstring
- hive.stats.jdbcdriver

These are no longer relevant after HIVE-12164 ("Remove jdbc stats
collection mechanism") in Hive 2.0.

- hive.metastore.rawstore.impl

This has always defaulted to 'ObjectStore' in Hive, so there was no
reason to set it explicitly.

- test.log.dir
- test.src.dir

These were listed in the config in a commented-out section. These were
commented out ever since 2012 when the file was first introduced.

This also fixes the postgres URL to not include a misplaced ';create'
parameter (which applies to Derby but not postgres).

Change-Id: Ief4434d80baae0fd7be7ffe7b2e07bae1ac45e47
Reviewed-on: http://gerrit.cloudera.org:8080/12930
Reviewed-by: Fredy Wijaya <fwijaya@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>