IMPALA-8336: fix flaky ORC memory limit test
Reduces the mem_limit for the ORC version of the
test, which has proven to be flaky.

Testing:
Looped the test for a while locally without any failures.

I was unable to reproduce the failure seen on CentOS6 jenkins locally,
so we'll just try this tweak and see if it improves this. If not
I can look deeper into it.

Change-Id: Iddcc34183ddcc7c48489739269881bffb1dc85e7
Reviewed-on: http://gerrit.cloudera.org:8080/12961
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>