IMPALA-8516. Update maven for Jenkins builds
This changes Maven to download and install on both Ubuntu and Redhat for
the Jenkins builds (previously it was only installed on Redhat).

The version number is kept at 3.5.4 even though a newer release is
available upstream. The new release fails to build Impala due to an
XML-parsing bug causing it to fail to resolve the parquet pom [1]

This should hopefully address some of the hang issues we've seen
previously with the older version of Maven that shipped with the version
of Ubuntu we have on Ubuntu 16.04.

[1] https://github.com/codehaus-plexus/plexus-utils/issues/65

Change-Id: I793409eb4e9f4533b75bfe089a497c0ea62ad1ff
Reviewed-on: http://gerrit.cloudera.org:8080/13268
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>
(cherry picked from commit 327b93821453f6abda2cb2d4437bba10946acc00)

Change-Id: Iac196e649a7f04bd86b4c1d792cc09757f6ea317
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/45909
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: quasar-precommit User
Quasar-L0: Jenkins User <jenkins@cloudera.com>