IMPALA-8491: Non-root user in container
Set a default USER in the Dockerfile per best practices so that
consumers of the container don't accidentally run as root.
The default user is "impala" if the container is run in docker
without specifying a user.

Various frameworks, including kubernetes, will run the container with
an arbitrary user and group ID set.

This causes issues with some Hadoop libraries, which depend on the
user having a name. This is generally not the case because inside
the container usernames are resolved with the container's /etc/passwd.

To work around this, the entrypoint script checks if the current
user has a name and if not, assigns it one (either dummyuser or
$HADOOP_USER_NAME).

Remove the umask setting that was required to make logs modifiable
by the host user - this is not needed for our tests since the host
host and container users now match up.

Also run apt-get clean in Dockerfile to reduce cruft in the
image.

Change-Id: I0bea9f44a8199851ed04fbef8caf4a2350ae2c0e
Reviewed-on: http://gerrit.cloudera.org:8080/13451
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 8cfd18ae89683869392693541f7ba05c1013076f)

Change-Id: Idbc04fddb9c2a3850b1fc4a8a3ad4e0925aeb65d
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/47761
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>