IMPALA-8443: Record time spent in authorization in the runtime profile
The analysis and authorization is handled together as authorization
depends on the results of the analysis. The timeline EventSequence is
moved into the AuthorizationContext and the markEvent is called during
the postAuthorize method.

In some cases the EventSequence is not available when the
AuthorizationContext is created therefore it is wrapped in Optional.

Change-Id: I5bb85e57fcc75d41f3eb2911e6d375e0da6f82ae
Reviewed-on: http://gerrit.cloudera.org:8080/13353
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 65175a23b2ed73d91734607d474ba80e15323c9b)

Change-Id: Ib414db14f5496ddb8addc49ad50a15bf63ae5d04
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/49712
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>