IMPALA-8486: fix stale libCache entries in LocalCatalog mode coordinators
In LocalCatalog mode, after a function is dropped, statestored will
broadcast the update to invalidate the cached CatalogObject in each
coordinator (if they have). However, the current code path does not
trigger libCache to remove the cached JAR/SO file. If we replace the
function file in HDFS with a new one and create the function again
using the same HDFS path, the SELECT statements in other coordinators
won't trigger libCache to refresh the local cached file, so they still
use the old cached file which causes errors.

When a coordinator invalidates its cached CatalogObject of a function,
it should also mark the corresponding libCache entry as "needs refresh".
So the later usage of this function will check the last modified time of
the HDFS file and refresh it in needs. To achieve this, we have to
propagate the HDFS path of the function along with the full function
name in the minimal topic, so libCache can target the cached entry.

Note that this does not prevent the dedicated executors to have stale
libCache entries. It needs some architecture changes. We'll follow it
in IMPALA-8763.

Tests
 - Re-enable test_udf_update_via_drop and test_udf_update_via_create for
LocalCatalog mode.

Change-Id: Ie4812fb8737de3ba6074ffeb9007927bfbbbaf9b
Reviewed-on: http://gerrit.cloudera.org:8080/13849
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 1cd85d1f8a0d772a4cab263cce4f41728f6ebac7)

Change-Id: Idee8a7bf8ee1a72456b9a137cd6f1997bbd9151a
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/51646
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>