IMPALA-8622,IMPALA-8696: fix docker dependencies, add image list
Adds a plain-text space-separated image list in
docker/docker-images.txt. This is generated based on the images built by
CMake, so is kept in sync with images added to or removed from the
CMake file.

Duplicated logic per image is removed - instead there is a helper
function that is called for each daemon image to be built.

Rips out the timestamp mechanism that was intended to avoid unnecessary
container rebuilds, but has turned out to be brittle. Instead the
containers are rebuilt each time the rule is invoked.

This moves some subdirectories so that the image tag matches the
subdirectory, to simplify the build scripts.

Change-Id: I4d8e215e9b07c6491faa4751969a30f0ed373fe3
Reviewed-on: http://gerrit.cloudera.org:8080/13899
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Lars Volker <lv@cloudera.com>
(cherry picked from commit f689daef7f140fb2e227ee96218f31076b5ba257)

Change-Id: I86aa2d7f1eefb1af5b5a898a15d663c5018b3543
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/52401
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>